backend/pyproject.toml
[tool.poetry]
name = "hotturbo-api"
version = "0.1.0"
description = "Backend API for hotturbo template repo"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "hotturbo_api"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.6.2"}
sqlmodel = "^0.0.16"
psycopg = {extras = ["binary"], version = "^3.1.18"}
alembic = "^1.13.1"
bcrypt = "^4.1.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
